+++ api/regen/ecocredit/v1/query_grpc.pb.go
@@ -93,6 +93,11 @@ type QueryClient interface {
 	//
 	// Since Revision 1
 	CreditClassFees(ctx context.Context, in *QueryCreditClassFeesRequest, opts ...grpc.CallOption) (*QueryCreditClassFeesResponse, error)
+	// AllowedBridgeChains queries the list of chains allowed to be used in bridge
+	// operations.
+	//
+	// Since Revision 1
+	AllowedBridgeChains(ctx context.Context, in *QueryAllowedBridgeChainsRequest, opts ...grpc.CallOption) (*QueryAllowedBridgeChainsResponse, error)
 }
 
 type queryClient struct {
@@ -329,6 +334,15 @@ func (c *queryClient) CreditClassFees(ctx context.Context, in *QueryCreditClassF
 	return out, nil
 }
 
+func (c *queryClient) AllowedBridgeChains(ctx context.Context, in *QueryAllowedBridgeChainsRequest, opts ...grpc.CallOption) (*QueryAllowedBridgeChainsResponse, error) {
+	out := new(QueryAllowedBridgeChainsResponse)
+	err := c.cc.Invoke(ctx, "/regen.ecocredit.v1.Query/AllowedBridgeChains", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 // QueryServer is the server API for Query service.
 // All implementations must embed UnimplementedQueryServer
 // for forward compatibility
@@ -404,6 +418,11 @@ type QueryServer interface {
 	//
 	// Since Revision 1
 	CreditClassFees(context.Context, *QueryCreditClassFeesRequest) (*QueryCreditClassFeesResponse, error)
+	// AllowedBridgeChains queries the list of chains allowed to be used in bridge
+	// operations.
+	//
+	// Since Revision 1
+	AllowedBridgeChains(context.Context, *QueryAllowedBridgeChainsRequest) (*QueryAllowedBridgeChainsResponse, error)
 	mustEmbedUnimplementedQueryServer()
 }
 
@@ -485,6 +504,9 @@ func (UnimplementedQueryServer) AllowedClassCreators(context.Context, *QueryAllo
 }
 func (UnimplementedQueryServer) CreditClassFees(context.Context, *QueryCreditClassFeesRequest) (*QueryCreditClassFeesResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreditClassFees not implemented")
+}
+func (UnimplementedQueryServer) AllowedBridgeChains(context.Context, *QueryAllowedBridgeChainsRequest) (*QueryAllowedBridgeChainsResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method AllowedBridgeChains not implemented")
 }
 func (UnimplementedQueryServer) mustEmbedUnimplementedQueryServer() {}
 
@@ -949,6 +971,24 @@ func _Query_CreditClassFees_Handler(srv interface{}, ctx context.Context, dec fu
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Query_AllowedBridgeChains_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(QueryAllowedBridgeChainsRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(QueryServer).AllowedBridgeChains(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/regen.ecocredit.v1.Query/AllowedBridgeChains",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(QueryServer).AllowedBridgeChains(ctx, req.(*QueryAllowedBridgeChainsRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 // Query_ServiceDesc is the grpc.ServiceDesc for Query service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -1055,6 +1095,10 @@ var Query_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "CreditClassFees",
 			Handler:    _Query_CreditClassFees_Handler,
+		},
+		{
+			MethodName: "AllowedBridgeChains",
+			Handler:    _Query_AllowedBridgeChains_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},