+++ x/ecocredit/base/types/v1/msg_add_allowed_bridge_chain.go
@@ -1,0 +1,41 @@
+package v1
+
+import (
+	"cosmossdk.io/errors"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
+
+	"github.com/regen-network/regen-ledger/x/ecocredit"
+)
+
+var _ legacytx.LegacyMsg = &MsgAddAllowedBridgeChain{}
+
+// Route implements the LegacyMsg interface.
+func (m MsgAddAllowedBridgeChain) Route() string { return sdk.MsgTypeURL(&m) }
+
+// Type implements the LegacyMsg interface.
+func (m MsgAddAllowedBridgeChain) Type() string { return sdk.MsgTypeURL(&m) }
+
+// GetSignBytes implements the LegacyMsg interface.
+func (m MsgAddAllowedBridgeChain) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ecocredit.ModuleCdc.MustMarshalJSON(&m))
+}
+
+// ValidateBasic does a sanity check on the provided data.
+func (m *MsgAddAllowedBridgeChain) ValidateBasic() error {
+	if _, err := sdk.AccAddressFromBech32(m.Authority); err != nil {
+		return errors.Wrapf(err, "invalid authority address")
+	}
+	if m.ChainName == "" {
+		return sdkerrors.ErrInvalidRequest.Wrap("chain_name cannot be empty")
+	}
+	return nil
+}
+
+// GetSigners returns the expected signers for MsgAddAllowedBridgeChain.
+func (m *MsgAddAllowedBridgeChain) GetSigners() []sdk.AccAddress {
+	addr, _ := sdk.AccAddressFromBech32(m.Authority)
+	return []sdk.AccAddress{addr}
+}